Refactor FileHandler Class
